tests/requirements.txt
pytest==3.6.2
docker==3.4.0
docker-compose==1..21.1
locustio==0.8.1
